fix: add missing imports
