+++ src/App.tsx
@@ -2,7 +2,7 @@ import { Route, Routes } from "react-router"
 import { EditProfile, Home } from "@src/pages"
 import Layout from "@src/layout/Layout"
 import PrivateLayout from "@src/layout/PrivateLayout"
-import AuthProvider from "@src/context/AuthContext"
+import AuthProvider from "@src/context/auth/AuthProvider"
 
 function App() {
