+++ src/context/auth/auth.module.ts
@@ -1,0 +1,17 @@
+import { TAuth } from "@src/types";
+import { createContext, useContext } from "react";
+
+export const AuthContext = createContext<TAuth>({
+	login: async () => { },  // Empty function to avoid undefined
+	logout: () => { },
+	session: null,
+});
+
+
+export const useAuth = () => {
+	const context = useContext(AuthContext);
+	if (context === undefined) {
+		throw new Error("useAuth must be used within an AuthProvider");
+	}
+	return context;
+};
